Move token verifier out of mutation
